Provide a parametric combinator that is atLeast but preserves owned configs (#98)
This introduces `parametric.withOwn` that is just a `parametric.atLeast`
that also includes the aspect owned configs.

See discussion:
https://github.com/vic/den/discussions/91#discussioncomment-14985716

Calling `parametric x` itself is an alias for `parametric.withOwn x`. 

So that both syntax do the same:

```
den.aspects.foo = {
  __functor = den.lib.parametric;
  nixos.foo = 1;
  includes = [ bar ];
};
```

and also, **this is the preferred* syntax that will be documented, since
it is more intuitive and does not surfaces `__functor`:

```

den.aspects.foo = den.lib.parametric {
  nixos.foo = 1;
  includes = [ bar ];
};
```

Still have to update documentation about this combinator.

Fixes #97.